+++ src/main/java/ladder/utils/LineStrategyRandom.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.utils;
 
 
 import java.util.Random;
