+++ src/test/java/ladder/present/RendererTest.java
@@ -5,7 +5,10 @@ import ladder.domain.Line;
 import ladder.domain.Scene;
 import ladder.domain.User;
 import ladder.domain.Users;
+import ladder.utils.LineStrategy;
+import ladder.utils.LineStrategyRandom;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -59,5 +62,22 @@ class RendererTest {
                             .isEqualTo(actual.get(index));
                 })
         );
+    }
+
+    public static class LineStrategyRandomTest {
+
+        private final LineStrategy lineStrategy = LineStrategyRandom.ofLimit(10, 555);
+
+        @DisplayName("지정한 Limit 범위 내로 값이 생성된다")
+        @RepeatedTest(300)
+        public void randomRange() {
+            //given
+            //when
+            //then
+            assertAll("랜덤값 생성을 검증한다",
+                    () -> assertThat(lineStrategy.anyColumn()).as("범위 내에서만 Column 랜덤 값을 생성한다").isBetween(0, 10),
+                    () -> assertThat(lineStrategy.anyRow()).as("범위 내에서만 Row 랜덤 값을 생성한다").isBetween(0, 555)
+            );
+        }
     }
 }
